Chapter 09/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Leonardo Lazzaro <llazzaro@dc.uba.ar>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
psycopg = "^3.1.9"
django-extensions = "^3.2.3"
django-storages = "^1.14"
django-widget-tweaks = "^1.5.0"
django-redis = "^5.3.0"
pillow = "^10.0.1"
django-ninja = "1.0b2"
pyjwt = "^2.8.0"
django-ratelimit = "^4.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
